add to compare test convert to junit
